+++ example/ffinative/lib/generated_bindings.dart
@@ -7,28 +7,35 @@
 import 'dart:ffi' as ffi;
 
 /// Adds 2 integers.
-@ffi.Native<ffi.Int Function(ffi.Int, ffi.Int)>(symbol: 'sum')
+@ffi.Native<ffi.Int Function(ffi.Int, ffi.Int)>(
+    symbol: 'sum', assetId: 'package:ffinative_example/generated_bindings.dart')
 external int sum(
   int a,
   int b,
 );
 
 /// Subtracts 2 integers.
-@ffi.Native<ffi.Int Function(ffi.Int, ffi.Int)>(symbol: 'subtract')
+@ffi.Native<ffi.Int Function(ffi.Int, ffi.Int)>(
+    symbol: 'subtract',
+    assetId: 'package:ffinative_example/generated_bindings.dart')
 external int subtract(
   int a,
   int b,
 );
 
 /// Multiplies 2 integers, returns pointer to an integer,.
-@ffi.Native<ffi.Pointer<ffi.Int> Function(ffi.Int, ffi.Int)>(symbol: 'multiply')
+@ffi.Native<ffi.Pointer<ffi.Int> Function(ffi.Int, ffi.Int)>(
+    symbol: 'multiply',
+    assetId: 'package:ffinative_example/generated_bindings.dart')
 external ffi.Pointer<ffi.Int> multiply(
   int a,
   int b,
 );
 
 /// Divides 2 integers, returns pointer to a float.
-@ffi.Native<ffi.Pointer<ffi.Float> Function(ffi.Int, ffi.Int)>(symbol: 'divide')
+@ffi.Native<ffi.Pointer<ffi.Float> Function(ffi.Int, ffi.Int)>(
+    symbol: 'divide',
+    assetId: 'package:ffinative_example/generated_bindings.dart')
 external ffi.Pointer<ffi.Float> divide(
   int a,
   int b,
@@ -36,7 +43,8 @@ external ffi.Pointer<ffi.Float> divide(
 
 /// Divides 2 floats, returns a pointer to double.
 @ffi.Native<ffi.Pointer<ffi.Double> Function(ffi.Float, ffi.Float)>(
-    symbol: 'dividePrecision')
+    symbol: 'dividePrecision',
+    assetId: 'package:ffinative_example/generated_bindings.dart')
 external ffi.Pointer<ffi.Double> dividePrecision(
   double a,
   double b,