+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   generate a typedef for the `Function`.
 - Use Dart wrapper types in args and returns of ObjCBlocks.
 - Bump min SDK version to 3.2.0-114.0.dev.
+- Renamed `asset` to `assetId` for `ffi-native`.  
 
 # 9.0.1